fix: change navigation
